requirements.txt
langchain
langchain-openai
langchain-community
langchain-experimental
langchain-text-splitters
langgraph
langsmith
openai
pandas
faiss-cpu
rank_bm25
tiktoken
python-dotenv
